+++ packages/lib/test/model/setterAction.test.ts
@@ -5,10 +5,10 @@ import { autoDispose } from "../utils"
 
 @model("P")
 export class P extends Model({
-  y: prop(0, { setterAction: true }),
-  z: tProp(0, { setterAction: true }),
-  sy: prop(0, { setterAction: "assign" }),
-  sz: tProp(0, { setterAction: "assign" }),
+  y: prop(0, { setter: true }),
+  z: tProp(0, { setter: true }),
+  sy: prop(0, { setter: "assign" }),
+  sz: tProp(0, { setter: "assign" }),
 }) {
   @computed
   get cy() {
@@ -31,7 +31,7 @@ export class P extends Model({
   }
 }
 
-test("setterAction", () => {
+test("setter", () => {
   const events: any = []
 
   const p = new P({})